register backends in strategy method
