+++ pandera/api/dataframe/container.py
@@ -20,7 +20,6 @@ from typing import (
 )
 
 from pandera import errors
-from pandera.import_utils import strategy_import_error
 from pandera.api.base.schema import BaseSchema, inferred_schema_guard
 from pandera.api.base.types import CheckList, ParserList, StrictType
 from pandera.api.checks import Check
@@ -1271,52 +1270,6 @@ class DataFrameSchema(Generic[TDataObject], BaseSchema):
         import pandera.io
 
         return pandera.io.to_json(self, target, **kwargs)
-
-    ###########################
-    # Schema Strategy Methods #
-    ###########################
-
-    @strategy_import_error
-    def strategy(
-        self, *, size: Optional[int] = None, n_regex_columns: int = 1
-    ):
-        """Create a ``hypothesis`` strategy for generating a DataFrame.
-
-        :param size: number of elements to generate
-        :param n_regex_columns: number of regex columns to generate.
-        :returns: a strategy that generates pandas DataFrame objects.
-        """
-        from pandera import strategies as st
-
-        return st.dataframe_strategy(
-            self.dtype,
-            columns=self.columns,
-            checks=self.checks,
-            unique=self.unique,
-            index=self.index,
-            size=size,
-            n_regex_columns=n_regex_columns,
-        )
-
-    def example(
-        self, size: Optional[int] = None, n_regex_columns: int = 1
-    ) -> TDataObject:
-        """Generate an example of a particular size.
-
-        :param size: number of elements in the generated DataFrame.
-        :returns: pandas DataFrame object.
-        """
-        # pylint: disable=import-outside-toplevel,cyclic-import,import-error
-        import hypothesis
-
-        with warnings.catch_warnings():
-            warnings.simplefilter(
-                "ignore",
-                category=hypothesis.errors.NonInteractiveExampleWarning,
-            )
-            return self.strategy(
-                size=size, n_regex_columns=n_regex_columns
-            ).example()
 
 
 def _validate_columns(
