+++ pandera/api/polars/components.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import logging
-from typing import Any, Optional
+from typing import Any, Optional, Type
 
 import polars as pl
 
@@ -107,7 +107,7 @@ class Column(ComponentSchema[PolarsCheckObjects]):
 
     @staticmethod
     def register_default_backends(
-        check_obj,
+        check_obj_cls: Type,
     ):  # pylint: disable=unused-argument
         register_polars_backends()
